Implementata classe carrello
